Async loading of chains
